Move appendToCollection from CLI to shared array utilities
Relocated appendToCollection function from src/cli/src/shared/collection.js
to src/shared/src/utils/array.js where it belongs with other array utilities.

Original location: src/cli/src/shared/collection.js
New location: src/shared/src/utils/array.js

Rationale:
- appendToCollection is a generic array utility, not CLI-specific
- It normalizes array/scalar accumulation, similar to toArray and asArray
- Moving it to shared/utils/array.js makes it available to all packages
- Co-locates it with related array manipulation functions
- Maintains all inline documentation and function signature

Changes:
- Added appendToCollection to src/shared/src/utils/array.js
- Updated src/cli/src/dependencies.js to import from @prettier-plugin-gml/shared
- Removed empty src/cli/src/shared/collection.js file
- All tests pass with no regressions
